+++ lib/pages/homepage.dart
@@ -83,7 +83,7 @@ class _HomePageState extends State<HomePage> {
 
         for (var item in data) {
           String name = item['name'];
-          int fundraiserBox = item['fundraiserBox'];
+          double fundraiserBox = item['fundraiserBox'];
           String? companyName;
 
           if (item['company'] != null) {
